hdcms录播视频提交

+++ hdcmslearn/system/model/Modules.php
@@ -21,6 +21,7 @@ use app\system\controller\part\Subscribe;
 use app\system\controller\part\Tag;
 use houdunwang\cli\Cli;
 use houdunwang\container\Container;
+use houdunwang\db\Db;
 use houdunwang\dir\Dir;
 use Request;
 
@@ -128,6 +129,7 @@ class Modules extends Common
      */
     protected static function defindConst()
     {
+        define('MODULE_VERSION', v('module.version'));
         define('MODULE_PATH', v('module.path'));
         define("MODULE_TEMPLATE_PATH", v('module.path')."/template");
         define("MODULE_TEMPLATE_URL", root_url().'/'.v('module.path')."/template");
@@ -844,12 +846,14 @@ class Modules extends Common
         rsort($data);
         foreach ($data as $file) {
             $info = pathinfo($file);
-            require $file;
-            $namespace = "addons\\{$name}\\database\migrations";
-            $class     = $namespace.'\\'.substr($info['basename'], 13, -4);
-            (new $class)->down();
+            if (Db::table('migrations')->where('migration', $info['basename'])->get()) {
+                require $file;
+                $namespace = "addons\\{$name}\\database\migrations";
+                $class     = $namespace.'\\'.substr($info['basename'], 13, -4);
+                (new $class)->down();
 
-            return Db::table('migrations')->where('migration', $info['basename'])->delete();
+                return Db::table('migrations')->where('migration', $info['basename'])->delete();
+            }
         }
     }
